parseChapter
